fix: optimize PO Dashboard performance with real contract integration
- Updated useProjectsByOwner to fetch real project modes from contract
- Implemented batch queries in useMultipleProjectDetails for better performance
- Fixed all TypeScript compilation errors (BigInt literals, import paths, type assertions)
- Removed 3-project limitation in project details fetching
- Updated contract addresses after redeployment with projectModes support
- Fixed array access issues in market maker hooks
- Added proper type assertions for wagmi ABI compatibility

The PO Dashboard now loads projects efficiently using the contract's getProjectMode
function instead of placeholder logic, resolving the "loading forever" issue.

+++ interface/components/bonding-curve/buy-sell-forms.tsx
@@ -31,10 +31,10 @@ export function BuySellForms({ bondingCurveAddress, usdcAddress, tokenSymbol }: 
     const calculateOutput = async () => {
       if (activeTab === 'buy' && buyAmount && parseFloat(buyAmount) > 0) {
         const tokens = await calculateBuyReturn(buyAmount);
-        setBuyTokensOut(tokens ? formatUnits(tokens, 18) : '0');
+        setBuyTokensOut(tokens ? formatUnits(tokens as bigint, 18) : '0');
       } else if (activeTab === 'sell' && sellAmount && parseFloat(sellAmount) > 0) {
         const usdc = await calculateSellReturn(sellAmount);
-        setSellUsdcOut(usdc ? formatUnits(usdc, 6) : '0');
+        setSellUsdcOut(usdc ? formatUnits(usdc as bigint, 6) : '0');
       }
     };
